Crate permissions

Permissions pallet provides an ability to configure an access via permissions.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Enums

Scope

Constants

BURN

CHECK_FARM

CLAIM_FROM_FARM

CREATE_FARM

GET_FARMER_INFO

GET_FARM_INFO

INIT_DEX

LOCK_TO_FARM

MANAGE_DEX

MINT

SLASH

UNLOCK_FROM_FARM

Type Definitions

HolderId

The id of the account having a permission

OwnerId

The id of the account owning a permission

PermissionId